refactor: move UserProfile and UserSocialMedia components into UserProfile own component folder
